Minor Enhancement
Paramount is great in practice and philosophy. I only had trouble with the purple being too dreamy/sleepy for programming and the hashed comments were very dimmed on my monitor I could not see them well. So I am just proposing a yellow-gold variation instead of the purple and a lighter grey for the comments section. I have tuned it for me and it's working, try it out and let me know.